obs: move silent-disconnect watchdog out of pkg/obs into pkg/obs/watchdog (#716)
The watchdog file imported pkg/config/tripbot (for ChannelName) and
pkg/twitch (for IsChannelLive at the package level — pkg/twitch in
turn imports pkg/oauthtokens which imports pkg/database, and that
init() log.Fatalf's on missing DATABASE_USER/DB/HOST). Since the
files all sat in package obs, anything that imported pkg/obs dragged
the whole chain in at link time.

vlc-server's main has imported pkg/obs for ages — for one call,
obs.PollStreamingActive. With the watchdog file in the same package,
vlc-server's binary inherited both the envconfig validator (6 keys)
and the database env-var check (3 keys). v2.17.0 (which added the
watchdog in #702) couldn't boot in prod without 9 placeholder env
vars stamped into vlc-server's ConfigMap — see adanalife/infra#617
and adanalife/infra#619 for the workaround that recovered prod.

Moving the watchdog into its own subpackage means pkg/obs holds only
control.go + streaming.go (zero tripbot-internal imports beyond
pkg/instrumentation), so any binary can import pkg/obs without
pulling in tripbot-specific machinery. cmd/tripbot is the sole
consumer of pkg/obs/watchdog.

Verified: `go list -deps ./cmd/vlc-server` no longer includes
pkg/config/tripbot, pkg/database, pkg/twitch, or pkg/oauthtokens.

After this lands, the placeholder env block can be removed from
infra's vlc-server base kustomization (paired PR).
